requirements.txt
streamlit==1.29.0
numpy==1.23.5
pandas==2.0.3
matplotlib==3.7.2
Pillow==10.0.0
opencv-python-headless==4.6.0.66
google-cloud-vision==3.10.1
google-cloud-translate==3.20.2
gTTS==2.3.2
torch==2.0.0
torchvision==0.15.1